Task 103/pvz tests

+++ pvz_service/src/services/pvz_groups_service.py
@@ -1,4 +1,4 @@
-from fastapi import HTTPException, status
+from sqlalchemy.exc import IntegrityError
 
 from src.dao.pvzGroupsDAO import PVZGroupsDAO
 from src.dao.pvzsDAO import PVZsDAO
@@ -7,6 +7,7 @@ from src.schemas.pvz_group_schemas import (
     PVZGroupResponseSchema,
     PVZGroupUpdateSchema,
 )
+from src.utils.exceptions import PVZGroupAlreadyExistsException, PVZGroupFilterException, PVZGroupNotFoundException
 
 
 class PVZGroupsService:
@@ -19,6 +20,10 @@ class PVZGroupsService:
         repo: PVZGroupsDAO,
     ):
         """Создаёт новую группу ПВЗ"""
+        existing_group = await repo.get_group(name=data.name, owner_id=data.owner_id)
+        if existing_group:
+            raise PVZGroupAlreadyExistsException("Группа с таким именем уже существет")
+
         payload = data.model_dump()
         group = await repo.create(payload)
 
@@ -28,14 +33,22 @@ class PVZGroupsService:
         self,
         group_id: int,
         data: PVZGroupUpdateSchema,
-        repo: PVZsDAO,
+        pvz_repo: PVZsDAO,
+        group_repo: PVZGroupsDAO,
     ):
         """Обновляет данные группы ПВЗ и кураторство для ПВЗ при необходимости."""
 
-        group = await repo.update(group_id, **data.model_dump(exclude_unset=True))
+        existing_group = await group_repo.get_by_id(id=group_id)
+        if not existing_group:
+            raise PVZGroupNotFoundException("Группы с таким id не существет")
+
+        try:
+            group = await group_repo.update(group_id, **data.model_dump(exclude_unset=True))
+        except IntegrityError:
+            raise PVZGroupAlreadyExistsException("Группа с таким именем уже существует") from None
 
         if data.curator_id:
-            await repo.update_pvzs_curator_by_group(group_id, data.curator_id)
+            await pvz_repo.update_pvzs_curator_by_group(group_id, data.curator_id)
 
         return PVZGroupResponseSchema.model_validate(group)
 
@@ -54,20 +67,14 @@ class PVZGroupsService:
             return [PVZGroupResponseSchema.model_validate(g) for g in groups]
 
         if len(actual_params) > 1:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Используйте только один параметр для поиска.",
-            )
+            raise PVZGroupFilterException("Используйте только один параметр для поиска.")
 
         field_name, field_value = next(iter(actual_params.items()))
 
         groups = await repo.get_groups(**{field_name: field_value})
 
         if not groups:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Группы не найдены",
-            )
+            raise PVZGroupNotFoundException("Группы не найдены")
 
         return [PVZGroupResponseSchema.model_validate(g) for g in groups]
 
@@ -81,7 +88,7 @@ class PVZGroupsService:
         group = await repo.get_by_id(group_id)
 
         if not group:
-            raise HTTPException(status.HTTP_404_NOT_FOUND, detail="Группа не найдена")
+            raise PVZGroupNotFoundException("Группа не найдена")
 
         return PVZGroupResponseSchema.model_validate(group)
 
@@ -92,10 +99,18 @@ class PVZGroupsService:
     ):
         """Удаляет группу ПВЗ."""
 
+        existing_group = await repo.get_by_id(id=group_id)
+        if not existing_group:
+            raise PVZGroupNotFoundException("Группы с таким id не существет")
+
         await repo.delete(id=group_id)
 
     async def assign_curator(self, group_id: int, curator_id: int, repo: PVZGroupsDAO, pvz_repo: PVZsDAO):
         """Привязывает куратора к группе, а также ко всем пвз состоящим в этой группе"""
+
+        existing_group = await repo.get_by_id(id=group_id)
+        if not existing_group:
+            raise PVZGroupNotFoundException("Группы с таким id не существет")
 
         async with self.db_helper.async_session_maker() as session:
             async with session.begin():  # ← одна транзакция
